+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - **[BC]** Drop support for PHP 7.1
 - **[BC]** Remove `PackageInfo` class
 - **[BC]** Rename `AnyComparableInterface` to `AnyComparable`
+- **[BC]** Rename `ExtendedComparableInterface` to `ExtendedComparable`
+- **[BC]** Remove `AbstractExtendedComparable` class, use `ExtendedComparableTrait` instead
 
 ### 2.0.0 (2018-11-06)